refactor: architecture and media type selection
