+++ README.md
@@ -29,13 +29,6 @@ Build image :
 
 For more, read the [FastAPI Dockerfile documentation](https://github.com/tiangolo/uvicorn-gunicorn-fastapi-docker#environment-variables)
 
-## Develop
-
-To run the app in debug mode :
-
-- Install deps : `pip3 install -r requirements.txt`
-- Run the run with hot reloading : `uvicorn main:app --reload`
-
 ### Debugging in VS Code
 
 You can debug the code by pressing F5 key.
