+++ TODO.md
@@ -1,7 +1,6 @@
 # TODO list
 
 - Make issues for TODOs
-- Refactor the [folder structure](https://fastapi.tiangolo.com/tutorial/bigger-applications/)
 - A user can download a playlist, probably a zip file
 - Host live on [Deta](https://fastapi.tiangolo.com/deployment/deta/)
 - Stream the file response while it is downloaded on the server, for shorten the download time
