+++ app/downloaders/download.py
@@ -1,6 +1,8 @@
 from __future__ import unicode_literals
 import os
 import youtube_dl
+from enum import Enum  
+
 
 title = ''
 
@@ -60,3 +62,16 @@ async def download(url: str, shouldDownloadPlaylist: bool, download_dir: str):
         ydl.download([url])
 
     return title
+
+class AudioFormat(Enum):
+    WAV = "audio/wav"
+    OGG = "audio/ogg"
+    WEBM = "audio/webm"
+
+def setAudioFormat(param: str):
+    switcher={
+        'wav':'audio/wav',
+        'ogg':'audio/ogg',
+        'webm':'audio/webm',
+        }
+    return switcher.get(param,"Invalid audio format")
