+++ app/main.py
@@ -1,13 +1,15 @@
 from typing import Optional
 
 from starlette.background import BackgroundTask
-from youtube.utils import slugify, delete_subdir_download
 from fastapi import FastAPI
 from fastapi.responses import FileResponse, RedirectResponse
-from youtube.download import download
 import os
+from downloaders import download
+from filesManagers.utils import slugify, delete_subdir_download
 
 DOWNLOAD_DIR = os.getenv('DOWNLOAD_DIR', "./downloaded/")
+
+
 
 app = FastAPI()
 
@@ -20,7 +22,7 @@ async def read_root():
 
 
 @app.get("/download")
-async def download_audio(url: Optional[str] = None, playlist: Optional[bool] = False):
+async def download_audio(url: Optional[str] = None, playlist: Optional[bool] = False, media_type: Optional[str] = 'webm'):
     current_download_dir = os.path.join(DOWNLOAD_DIR, slugify(url))
     pathfile = await download(url, playlist, current_download_dir)
     filename = pathfile.replace(current_download_dir, "")
@@ -28,13 +30,14 @@ async def download_audio(url: Optional[str] = None, playlist: Optional[bool] = F
     # getting the path only
     subdir_path = pathfile.replace(filename, "")
     clean_filename = filename.replace("NA -", "")
+    MEDIA_TYPE = setAudioFormat(media_type)
 
     # audio media_type info :
     # https://developer.mozilla.org/en-US/docs/Web/HTTP/Basics_of_HTTP/MIME_types#audio_and_video_types
     return FileResponse(
         path=pathfile,
         filename=clean_filename,
-        media_type="audio/webm",
+        media_type=MEDIA_TYPE,
         background=BackgroundTask(
             delete_subdir_download, path=subdir_path, wait_minute=1)
     )
